requirements.txt
astroid==3.3.8
asttokens==3.0.0
black==25.1.0
click==8.1.8
comm==0.2.2
debugpy==1.8.12
decorator==5.2.1
dill==0.3.9
executing==2.2.0
Faker==36.1.1
greenlet==3.1.1
ipykernel==6.29.5
ipython==8.32.0
isort==6.0.1
jedi==0.19.2
jupyter_client==8.6.3
jupyter_core==5.7.2
loguru==0.7.3
matplotlib-inline==0.1.7
mccabe==0.7.0
mypy-extensions==1.0.0
nest-asyncio==1.6.0
numpy==2.2.3
packaging==24.2
pandas==2.2.3
parso==0.8.4
pathspec==0.12.1
pexpect==4.9.0
platformdirs==4.3.6
prompt_toolkit==3.0.50
psutil==7.0.0
psycopg2-binary==2.9.10
ptyprocess==0.7.0
pure_eval==0.2.3
Pygments==2.19.1
PyLink==0.3.3
pylint==3.3.4
pyserial==3.5
python-dateutil==2.9.0.post0
pytz==2025.1
PyYAML==6.0.2
pyzmq==26.2.1
six==1.17.0
SQLAlchemy==1.4.54
stack-data==0.6.3
tomlkit==0.13.2
tornado==6.4.2
traitlets==5.14.3
typing_extensions==4.12.2
tzdata==2025.1
wcwidth==0.2.13
